docs: added comments to code
